3_2_redes2/practica3/requirements.txt
appJar==0.94.0
numpy==1.20.3
opencv-python==4.5.2.52
Pillow==8.2.0
